requirements.txt
fastapi
uvicorn[standard]
openai
motor
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
PyPDF2
pydantic-settings
